Disable clap dependency for bindgen
Since 0.50.0, bindgen is buildable without clap. clap is a command line application
framework that is used for the bindgen cli program. onig_sys uses bindgen's programmatic
interface in build.rs, so this dependency is not needed.

Disabling the clap feature of bindgen brings the total deps required to build onig
from 41 to 30.